Add files via upload
Facebook Data Visualisation uses EDA and Data Profiling to understand customer usage patterns of social media. 
This includes the below observations:
The spread of data for females is between 20 and 55 years of age and for males it is between 20 to 45 years of age. 
Younger males seem to be more engaged hence further analysis can be done to identify their behavioural patterns and the content they engage with , implying higher social media presence.
Popular males and females belonging to a certain age group ( 18-22 ) approximately have the highest number of likes received. Hence , closely analysing their social media presence with time of the day, the different pages they have created, their personalities etc can be tracked and social media analytics can play a vital role in it.
Usage of mobile device to access Facebook is high for people in the age group of 18 to 30. However, for high number of likes given through the web, seems to be spread till the age of 65-70 years. 
In order to encourage users to interact with Facebook more for social media marketing, it is important to  gather further data on small and large business pages on Facebook and how different individuals are leveraging Facebook for it.